# mypostbank-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI specification - Yasin",
    "description": "OpenAPI Documentation for MyPostBank",
    "termsOfService": "Terms of Service",
    "contact": {
      "name": "Yasin",
      "email": "yasinjb002@gmail.com"
    },
    "license": {
      "name": "License"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/mypostbank",
      "description": "Local ENV"
    },
    {
      "url": "https://yasin",
      "description": "PROV ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/user/cards/{cardId}/deactivate": {
      "put": {
        "tags": [
          "card-controller"
        ],
        "summary": "Deactivate card by ID",
        "operationId": "deactivateCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/cards/{cardId}/activate": {
      "put": {
        "tags": [
          "card-controller"
        ],
        "summary": "Activate card by ID",
        "operationId": "activateCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/accounts/{accountId}/activate": {
      "put": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Activate account by ID",
        "operationId": "activateAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/credit-requests/{creditId}/reject": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Reject credit request by ID",
        "operationId": "rejectCredit",
        "parameters": [
          {
            "name": "creditId",
            "in": "path",
            "description": "Credit request ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credit request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Credit request not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cards/{cardId}/reject": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Reject card by ID",
        "operationId": "rejectCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cards/{cardId}/deactivate": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Deactivate card by ID",
        "operationId": "deactivateCard_1",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cards/{cardId}/approve": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Approve card by ID",
        "operationId": "approveCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cards/{cardId}/activate": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Activate card by ID",
        "operationId": "activateCard_1",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/accounts/{accountId}/freeze": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Freeze account by ID",
        "operationId": "freezeAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/accounts/{accountId}/deactivate": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Reject account by ID",
        "operationId": "rejectAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/credit-requests/{creditId}/approve": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Approve credit request by ID",
        "operationId": "approveCredit",
        "parameters": [
          {
            "name": "creditId",
            "in": "path",
            "description": "Credit request ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credit request approved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Credit request not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/checkbook/{checkId}/update/{checkbookStatus}": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "summary": "update checkbook status",
        "operationId": "updateCheckbookStatus",
        "parameters": [
          {
            "name": "checkId",
            "in": "path",
            "description": "checkbook id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "checkbookStatus",
            "in": "path",
            "description": "status",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ORDERED",
                "BEING_DELIVERED",
                "RECEIVED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/accounts/{accountId}/activate": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Activate account by ID",
        "operationId": "activateAccount_1",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{userId}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get user by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/admin/transactions": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get all transactions",
        "operationId": "getAllTransactions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseTransactionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/admin/transactions/{transactionId}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get transaction by ID",
        "operationId": "getTransaction",
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "description": "Transaction ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/loans/{creditId}/installments": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get credit installments",
        "operationId": "getInstallments",
        "parameters": [
          {
            "name": "creditId",
            "in": "path",
            "description": "Credit request ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepaymentDetailDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/accounts/{accountId}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get account by ID",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/cards/{cardId}/delete": {
      "delete": {
        "tags": [
          "card-controller"
        ],
        "summary": "delete card",
        "operationId": "deleteCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "query",
            "description": "card id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{userId}/delete": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Delete user by ID",
        "operationId": "deleteUser_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cards/{cardId}/delete": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Delete card by ID",
        "operationId": "deleteCard_1",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/accounts/{accountId}/close": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Close account by ID",
        "operationId": "closeAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "Account ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PageResponseTransactionDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDTO"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "TransactionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "TRANSFER",
              "CARD_PAYMENT",
              "WITHDRAWAL",
              "DEPOSIT",
              "CREDIT_INSTALLMENT"
            ]
          },
          "amount": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "beneficiary": {
            "type": "integer",
            "format": "int32"
          },
          "cardId": {
            "type": "integer",
            "format": "int32"
          },
          "accountId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RepaymentDetailDTO": {
        "type": "object",
        "properties": {
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "principalPayment": {
            "type": "number"
          },
          "interestPayment": {
            "type": "number"
          },
          "remainingBalance": {
            "type": "number"
          },
          "paymentAmount": {
            "type": "number"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PAID",
              "NOT_PAID",
              "PARTIAL",
              "OVERDUE",
              "CANCELLED",
              "INSUFFICIENT_FUNDS"
            ]
          }
        }
      },
...
      "BankAccountDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "accountNo": {
            "type": "string"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "CHECKING",
              "SAVINGS"
            ]
          },
          "balance": {
            "type": "number"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PENDING",
              "INACTIVE",
              "FROZEN"
            ]
          },
          "owner": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "cinFile": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MANAGER"
            ]
          },
          "dateJoined": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}